migrate時にエラーを回避できるようにした。

+++ database/migrations/2024_04_05_020810_create_data_admins_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('data_admins', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('admin_id');
-            $table->string('permission_code', 255)->nullable();
-            $table->string('first_name', 255);
-            $table->string('last_name', 255);
-            $table->string('email', 255);
-            $table->timestamp('created_at')->useCurrent();
-            $table->timestamp('updated_at')->useCurrentOnUpdate();
-            $table->timestamp('deleted_at')->nullable();
+        if (!Schema::hasTable('data_admins')) {
+            Schema::create('data_admins', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('admin_id');
+                $table->string('permission_code', 255)->nullable();
+                $table->string('first_name', 255);
+                $table->string('last_name', 255);
+                $table->string('email', 255);
+                $table->timestamp('created_at')->useCurrent();
+                $table->timestamp('updated_at')->useCurrentOnUpdate();
+                $table->timestamp('deleted_at')->nullable();
 
-            $table->foreign('admin_id')->references('id')->on('master_admins')->onUpdate('CASCADE')->onDelete('CASCADE');
-        });
+                $table->foreign('admin_id')->references('id')->on('master_admins')->onUpdate('CASCADE')->onDelete('CASCADE');
+            });
+        }
     }
 
     /**
